+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- `hideOutOfStock` prop to `RelatedProducts` shelf
 ## [1.45.0] - 2021-06-09
 ### Added
 - List name to GTM `productClick` event.